+++ src/il/ac/hit/tasksmanager/model/Model.java
@@ -4,9 +4,9 @@ import il.ac.hit.tasksmanager.model.dao.ITasksDAO;
 import il.ac.hit.tasksmanager.model.dao.TasksDAOException;
 import il.ac.hit.tasksmanager.model.dao.TasksDAOProxy;
 import il.ac.hit.tasksmanager.model.dao.TasksDAOImpl;
-import il.ac.hit.tasksmanager.model.entities.TaskState;
-import il.ac.hit.tasksmanager.model.entities.ToDoState;
-import il.ac.hit.tasksmanager.model.patterns.TaskObserver;
+import il.ac.hit.tasksmanager.model.entities.state.TaskState;
+import il.ac.hit.tasksmanager.model.entities.state.ToDoState;
+import il.ac.hit.tasksmanager.model.observer.TaskObserver;
 
 import javax.swing.SwingUtilities;
 import java.time.LocalDate;
@@ -19,7 +19,12 @@ import java.util.concurrent.Executors;
 import java.util.function.Consumer;
 
 /**
- * Model implements IModel with a DAO backend and an executor for async operations.
+ * Model implements the application domain logic, backed by a DAO and
+ * an executor for asynchronous operations.
+ *
+ * /*
+ *  Observer Pattern: views subscribe via {@link #register(TaskObserver)} and
+ *  are notified on data changes using {@link #notifyObservers()}.
  */
 public class Model implements IModel {
 	private final ITasksDAO dao;
@@ -37,6 +42,9 @@ public class Model implements IModel {
 	}
 
 	@Override
+	/**
+	 * Loads tasks asynchronously from the DAO into the in-memory cache and notifies observers.
+	 */
 	public void loadData() {
 		executor.submit(() -> {
 			try {
@@ -48,6 +56,11 @@ public class Model implements IModel {
 		});
 	}
 
+	/**
+	 * Retrieves tasks asynchronously and invokes the given callback on the EDT.
+	 *
+	 * @param callback consumer that will receive the tasks array on the EDT
+	 */
 	public void getTasksAsync(Consumer<Task[]> callback) {
 		executor.submit(() -> {
 			try {
@@ -60,20 +73,34 @@ public class Model implements IModel {
 	}
 
 	@Override
+	/**
+	 * Persists data if needed. For embedded Derby every mutation is persisted immediately, so this is a no-op.
+	 */
 	public void saveData() {
 		// Derby is persisted on each mutation; method left for API completeness
 	}
 
 	@Override
+	/**
+	 * Returns an immutable snapshot of the currently cached tasks.
+	 *
+	 * @return unmodifiable list of tasks
+	 */
 	public List<Task> getTasks() {
 		return Collections.unmodifiableList(cached);
 	}
 
 	@Override
+	/**
+	 * Adds a basic task with default TODO state and no due date.
+	 *
+	 * @param title task title
+	 * @param description task description
+	 */
 	public void addTask(String title, String description) throws ModelException {
 		executor.submit(() -> {
 			try {
-				dao.addTask(new BasicTask(0L, title, description, new ToDoState(), null));
+				dao.addTask(new BasicTask(0, title, description, new ToDoState(), null));
 				cached = Arrays.asList(dao.getTasks());
 				notifyObservers();
 			} catch (TasksDAOException e) {
@@ -82,6 +109,15 @@ public class Model implements IModel {
 		});
 	}
 
+	/**
+	 * Adds a basic task with an explicit state and due date, validating the due date is not in the past.
+	 *
+	 * @param title task title
+	 * @param description task description
+	 * @param state desired state (TODO if null)
+	 * @param dueDate due date (nullable)
+	 * @throws ModelException if validation fails
+	 */
 	public void addTask(String title, String description, TaskState state, LocalDate dueDate) throws ModelException {
 		// Validate due date is not in the past
 		if (dueDate != null && dueDate.isBefore(java.time.LocalDate.now())) {
@@ -89,7 +125,7 @@ public class Model implements IModel {
 		}
 		executor.submit(() -> {
 			try {
-				dao.addTask(new BasicTask(0L, title, description, state == null ? new ToDoState() : state, dueDate));
+				dao.addTask(new BasicTask(0, title, description, state == null ? new ToDoState() : state, dueDate));
 				cached = Arrays.asList(dao.getTasks());
 				notifyObservers();
 			} catch (TasksDAOException e) {
@@ -98,6 +134,16 @@ public class Model implements IModel {
 		});
 	}
 
+	/**
+	 * Adds a recurring task with validation on interval and due date.
+	 *
+	 * @param title task title
+	 * @param description task description
+	 * @param state desired state (TODO if null)
+	 * @param dueDate due date (nullable)
+	 * @param recurrenceDays positive interval in days
+	 * @throws ModelException if validation fails
+	 */
 	public void addRecurringTask(String title, String description, TaskState state, LocalDate dueDate, int recurrenceDays) throws ModelException {
 		if (recurrenceDays <= 0) {
 			throw new ModelException("recurrenceDays must be positive");
@@ -107,7 +153,7 @@ public class Model implements IModel {
 		}
 		executor.submit(() -> {
 			try {
-				RecurringTask rt = new RecurringTask(0L, title, description, state == null ? new ToDoState() : state, dueDate, recurrenceDays);
+				RecurringTask rt = new RecurringTask(0, title, description, state == null ? new ToDoState() : state, dueDate, recurrenceDays);
 				dao.addTask(rt);
 				cached = Arrays.asList(dao.getTasks());
 				notifyObservers();
@@ -118,6 +164,11 @@ public class Model implements IModel {
 	}
 
 	@Override
+	/**
+	 * Updates an existing task using the DAO and refreshes the cache.
+	 *
+	 * @param task task with updated values
+	 */
 	public void updateTask(Task task) throws ModelException {
 		executor.submit(() -> {
 			try {
@@ -131,7 +182,12 @@ public class Model implements IModel {
 	}
 
 	@Override
-	public void deleteTask(long id) throws ModelException {
+	/**
+	 * Deletes a task by id using the DAO and refreshes the cache.
+	 *
+	 * @param id task identifier
+	 */
+	public void deleteTask(int id) throws ModelException {
 		executor.submit(() -> {
 			try {
 				dao.deleteTask(id);
@@ -144,15 +200,18 @@ public class Model implements IModel {
 	}
 
 	@Override
+	/** Registers a model observer that will be notified on data changes. */
 	public void register(TaskObserver observer) {
 		observers.add(observer);
 	}
 
 	@Override
+	/** Unregisters a previously registered observer. */
 	public void remove(TaskObserver observer) {
 		observers.remove(observer);
 	}
 
+	/** Notifies all registered observers about data changes (Observer pattern). */
 	private void notifyObservers() {
 		for (TaskObserver o : observers) {
 			o.onTasksChanged();
